extract TreeNode to PGN util to lib
